opt: split into 2 seperate file
